Converted addIngredient fragment to alertdialog
