+++ .idea/misc.xml
@@ -21,7 +21,7 @@
         <entry key="app/src/main/res/layout/activity_main.xml" value="0.2661900838216146" />
         <entry key="app/src/main/res/layout/activity_main2.xml" value="0.14235055724417425" />
         <entry key="app/src/main/res/layout/activity_register.xml" value="0.3296025594075521" />
-        <entry key="app/src/main/res/layout/add_ingredient_dialog.xml" value="0.7592762642166242" />
+        <entry key="app/src/main/res/layout/add_ingredient_dialog.xml" value="0.3006519966384992" />
         <entry key="app/src/main/res/layout/camera_dialog.xml" value="0.17015991210937506" />
         <entry key="app/src/main/res/layout/dialog.xml" value="0.34479166666666666" />
         <entry key="app/src/main/res/layout/filter_dialog.xml" value="0.44361572265625004" />
