+++ app/src/main/java/com/example/recipeapp/fragments/InventoryFragment.java
@@ -1,11 +1,13 @@
 package com.example.recipeapp.fragments;
 
+import android.app.ProgressDialog;
 import android.content.DialogInterface;
 import android.os.Bundle;
 import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+import android.widget.Toast;
 
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
@@ -18,6 +20,7 @@ import com.example.recipeapp.activities.MainActivity;
 import com.example.recipeapp.adapters.InventoryAdapter;
 import com.example.recipeapp.databinding.AddIngredientDialogBinding;
 import com.example.recipeapp.databinding.FragmentInventoryBinding;
+import com.example.recipeapp.models.Comment;
 import com.example.recipeapp.models.Ingredient;
 import com.example.recipeapp.models.User;
 import com.google.android.material.dialog.MaterialAlertDialogBuilder;
@@ -34,6 +37,7 @@ public class InventoryFragment extends Fragment {
     private List<Ingredient> ingredientList;
     private InventoryAdapter adapter;
     private User currentUser;
+    private ProgressDialog progressDialog;
 
     public InventoryFragment() {
 
@@ -43,8 +47,9 @@ public class InventoryFragment extends Fragment {
     public void onCreate(@Nullable Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         ingredientList = new ArrayList<>();
+        progressDialog = new ProgressDialog(getContext());
         adapter = new InventoryAdapter(getContext(), ingredientList);
-        ((MainActivity) getActivity()).getSupportActionBar().setTitle("Inventory");
+
     }
 
     @Override
@@ -56,7 +61,14 @@ public class InventoryFragment extends Fragment {
     }
 
     @Override
+    public void onDestroyView() {
+        super.onDestroyView();
+        ((MainActivity) getActivity()).getSupportActionBar().show();
+    }
+
+    @Override
     public void onViewCreated(@NonNull View view, @Nullable Bundle savedInstanceState) {
+        ((MainActivity) getActivity()).getSupportActionBar().hide();
         super.onViewCreated(view, savedInstanceState);
         LinearLayoutManager linearLayoutManager = new LinearLayoutManager(getContext());
         binding.rvIngredients.setAdapter(adapter);
@@ -78,6 +90,7 @@ public class InventoryFragment extends Fragment {
         } catch (ParseException e) {
             Log.e(TAG, "No user found");
         }
+        binding.tvNumIngredients.setText(ingredientList.size() + " items");
     }
 
     public void goToBarcodeScan() {
@@ -91,11 +104,45 @@ public class InventoryFragment extends Fragment {
         alertDialogBuilder.setPositiveButton("Add", new DialogInterface.OnClickListener() {
             @Override
             public void onClick(DialogInterface dialog, int which) {
-                // Other functions to add ingredient
+                progressDialog.show();
+                String name = ingredientDialogBinding.etName.getText().toString();
+                String count = ingredientDialogBinding.etCount.getText().toString();
+                String unit = ingredientDialogBinding.etUnit.getText().toString();
+                if (name.isEmpty() || count.isEmpty() || unit.isEmpty()) {
+                    Toast.makeText(getContext(), "Fields cannot be empty!", Toast.LENGTH_SHORT).show();
+                    return;
+                }
+                addIngredient(name, count, unit);
             }
         });
         alertDialogBuilder.setNegativeButton("Cancel", (dialog, which) -> dialog.dismiss());
         alertDialogBuilder.show();
-//        NavHostFragment.findNavController(this).navigate(R.id.addIngredientFragment);
     }
+
+    private void addIngredient(final String name, final String count, final String unit) {
+        Ingredient ingredient = new Ingredient();
+        ingredient.initialize(name, Integer.parseInt(count), unit);
+        ingredient.saveInBackground(e -> {
+            if (e != null) {
+                Log.e(TAG, "Error in adding ingredient!", e);
+                return;
+            }
+            List<Ingredient> ingredientList = currentUser.getIngredientArray();
+            adapter.clear();
+            ingredientList.add(ingredient);
+            adapter.addAll(ingredientList);
+            binding.tvNumIngredients.setText(ingredientList.size() + " items");
+            progressDialog.dismiss();
+            currentUser.setIngredientArray(ingredientList);
+            currentUser.getParseUser().saveInBackground(e1 -> {
+                if (e1 != null) {
+                    Log.e(TAG, "Error in adding ingredient to user!", e1);
+                    return;
+                }
+                Log.i(TAG, "Saved ingredient to user's ingredient list!");
+            });
+        });
+
+    }
+
 }
