+++ app/src/main/res/navigation/nav_graph.xml
@@ -63,11 +63,7 @@
         android:id="@+id/inventoryFragment"
         android:name="com.example.recipeapp.fragments.InventoryFragment"
         android:label="fragment_inventory"
-        tools:layout="@layout/fragment_inventory">
-        <action
-            android:id="@+id/action_inventoryFragment_to_addIngredientFragment"
-            app:destination="@id/addIngredientFragment" />
-    </fragment>
+        tools:layout="@layout/fragment_inventory"/>
     <fragment
         android:id="@+id/uploadRecipeFragment"
         android:name="com.example.recipeapp.fragments.UploadRecipeFragment"
@@ -78,15 +74,6 @@
         android:name="com.example.recipeapp.fragments.UploadPostFragment"
         android:label="fragment_upload_post"
         tools:layout="@layout/fragment_upload_post" />
-    <fragment
-        android:id="@+id/addIngredientFragment"
-        android:name="com.example.recipeapp.fragments.AddIngredientFragment"
-        android:label="fragment_add_ingredient"
-        tools:layout="@layout/fragment_add_ingredient">
-        <action
-            android:id="@+id/action_addIngredientFragment_to_inventoryFragment"
-            app:destination="@id/inventoryFragment" />
-    </fragment>
     <fragment
         android:id="@+id/barcodeScanFragment"
         android:name="com.example.recipeapp.fragments.BarcodeScanFragment"
